+++ src/services/SearchService.ts
@@ -1,9 +1,9 @@
-import { HttpClient, httpClient } from '../apis/HttpClient';
+import { HttpClient } from '../apis/HttpClient';
 /*
   get(endpoint,options)=> Promise<Search[]>
 */
 
-class SearchService {
+export class SearchService {
   private httpClient;
   constructor(httpClient: HttpClient) {
     this.httpClient = httpClient;
@@ -20,4 +20,3 @@ class SearchService {
 }
 
 //인스턴화 및 의존성 주입
-export const searchService = new SearchService(httpClient);
